+++ src/components/ReframeROI.tsx
@@ -139,7 +139,7 @@ export default class ReframeROI {
         this._overlay = new Overlay({
             className: `${CLASS_OVERLAY} ${mode}-mode`,
             element: this._rectangle,
-            stopEvent: false
+            stopEvent: true
         });
 
         this._map.addOverlay(this._overlay);
@@ -182,26 +182,33 @@ export default class ReframeROI {
     }
 
     private _getExtent(): Polygon {
-        const overlayFrame = this._rectangle.getBoundingClientRect();
+        const mapBounds = this._map.getTargetElement().getBoundingClientRect();
+
+        const overlayFrameBounds = this._rectangle.getBoundingClientRect();
+
+        const relativePosition = {
+            left: overlayFrameBounds.left - mapBounds.left,
+            top: overlayFrameBounds.top - mapBounds.top
+        };
 
         const topLeft = this._map.getCoordinateFromPixel([
-            overlayFrame.left,
-            overlayFrame.top
+            relativePosition.left,
+            relativePosition.top
         ]);
 
         const topRight = this._map.getCoordinateFromPixel([
-            overlayFrame.left + overlayFrame.width,
-            overlayFrame.top
+            relativePosition.left + overlayFrameBounds.width,
+            relativePosition.top
         ]);
 
         const bottomLeft = this._map.getCoordinateFromPixel([
-            overlayFrame.left,
-            overlayFrame.top + overlayFrame.height
+            relativePosition.left,
+            relativePosition.top + overlayFrameBounds.height
         ]);
 
         const bottomRight = this._map.getCoordinateFromPixel([
-            overlayFrame.left + overlayFrame.width,
-            overlayFrame.top + overlayFrame.height
+            relativePosition.left + overlayFrameBounds.width,
+            relativePosition.top + overlayFrameBounds.height
         ]);
 
         return new Polygon([[topLeft, topRight, bottomRight, bottomLeft]]);