+++ CHANGELOG.md
@@ -138,3 +138,8 @@
 
 ## v2.0.5
 * Restore contrain resolution view if reframe is cancelled
+
+## v2.0.6
+* Improved pointer events and stop event overlay
+* Fixed reframe position when map is not full height and width
+* Added some transparency to the reframe control buttons